pyproject.toml
[tool.poetry]
name = "cradle-app"
version = "0.1.1"
description = "A tool to test ML models"
authors = ["swiftdiaries <adhita94@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
cradle-app = "cradle_app.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.3.0"}
pre-commit = "^2.6.0"
black = "^19.10b0"
fastapi = {extras = ["all"], version = "^0.59.0"}
gradio = "^1.0.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
